Can't finish import wizard in some cases
On first page, Finish is enabled when page is complete and there
are no lifecycle mapping errors and no selected proposals.

On second page, Finish is enabled when there are no selected
proposals. Note it is possible to finish import even if there are
mapping problems.

Moved proposal autocomplete to a later moment (when mapping page
set visible), so if Finish is pressed on the first page the wizard
will not attempt to install anything.

Fixed number of issues with mapping proposals not showing on the
mapping page or showing with wrong icon and/or text.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>