342910 use 'never' global update policy by default
Introduced new "do not update dependency from remote repositories"
preference. When checked (the default) m2e will only contact remote
repositories to resolve artifacts that are not present in the local
repository and will never attempt to re-resolve artifacts that are
already present locally.

This is expected to guarantee that all workspace project use the same
SNAPSHOT dependency version and improve dependency resolution
performance in many common cases as well.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>