350956 Workspace resolution doesn't find newly opened projects
This is a regression introduced by performance optimizations done
for 342903. MavenBuilder was expected to synchronously update
project registry as part of a build when workspace was autobuilding,
but workspace does not build project after open events without other
changes.

Solution is to force background registry refresh for open project
events. This degrades performance and/or concurrency in the case
when workspace does run build after project open event, but project
close/open are relatively rare and overall performance import is
not expected to be noticeable.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>