350443 - Quickfix for lifecycle mapping ignore does not always work
Fixed a problem when quick fix dialog did not appear right after
workbench restart, when MavenProject cache was not primed yet. The
code now calls getMavenProject(IProgressMonitor), which loads
MavenProject instance as necessary.

Fixed a problem when quick-fix failed if plugin was using ${prop}
version. Unfortunately, the code will still fail if plugin groupId
or artifactId are defined using properties, but I to fix that
association between elements of effective and original project
model is necessary, but Maven does not currently provide it.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>