349586 introduced notion of pre-configuration build
Pre-configuration build is executed as part of project import
and update configuration and is meant to allow changes to the
mutatable state of MavenProject instance, i.e. sources roots,
resource and properties.

Pre-configuration buidl is executed with special "no-changes"
build context, which allows direct execution of maven plugins
compatible with workspace incremental build.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>