368230 : new API for Runtime classpath resolution
3rd party plugins can contribute IClassifierClasspathProvider
implementations via the new classifierClasspathProvider extension point.
These implementations should resolve the appropriate classpath (main
or test) for a given classifier during classpath resolution of Launch
configurations by the MavenRuntimeClasspathProvider.

For instance, for the ejb "client" classifier, a dedicated
classifierClasspathProvider extension would make available the EJB project
classes to the main classpath, but not its test classes to the test
classpath.

It's recommended the implementations extend the
AbstractClassifierClasspathProvider, as to provide better forward
compatibility.

Signed-off-by: Fred Bricon <fbricon@gmail.com>