This commit contains most of the work for bug 350414. This includes adding workspace settings to the chain of lookups for determining lifecycle mappings for m2e projects. There are changes to existing quickfixes, and a new workspace preferences page that allows users to view their workspace mappings and edit the raw text file.
There is partial work done for displaying a table of workspace
lifecycle mappings in the preferences page.  The refactoring
of the project lifecycle mappings table is complete, but I
just don't know how to populate the viewer with workspace
mappings only

For compliance with Eclipse.org's contribution rules:
(1) I (Andrew Eisenberg) wrote 100% of the code;
(2) I have the right to contribute the code to Eclipse;
(3) the file header contains the appropriate License header.

Also, this patch is built on top of the SHA 170c7936da947d28e7824bc1d2375903dbda9b17