394525 explicit configuration of project lifecycleMappingId
Introduced new ResolverConfiguration#lifecycleMappingId attribute.
If the attribute is not null, the ILifecycleMapping with specified
id will be used regardless of any other lifecycle mapping metadata
applicable to the project.

The new attribute is meant for m2e exploiters and is not and most
likely will not be exposed via user-accessible UI.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>