perf: don't refresh project facade during full build
Changed maven builder to refresh facade when project pom.xml has
changed or facade is not available or facade is stale. Use cached
facade instance in all other cases. This improves performance in
common case when project configuration update and clean build
are performed as one user operations.

May cause stale facade cache when full build is invoked right after
out-of-workspace pom.xml change. This is (and always was) an edge
case, if this happens, the workaround is to manually invoke project
configuration update.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>