433525 : Fix background archetype retrieval
Archetype retrieval job is now cancelled every time the catalog changes or
the user goes to the next page. The selected archetype variable is only
updated if the job finishes while on the same page, to avoid updating
disposed widgets or the current archetype while another page is displayed.

Bug 433525
Change-Id: Id78104ce2ec1aed0fd090220ee98cf1b941b5396
Signed-off-by: Fred Bricon <fbricon@gmail.com>