403243 prevent accidental wiring to m2e-bundled thirdparty libraries
m2e maven runtime and related bundles embed a number of thirdpaty
libraries that are meant to be used in conjunction with m2e and are
not general purpose providers of these libraries. To signify and
enforce such usage, annotate exported package definitions with
mandatory provider=m2e attribute. this prevents accidental wired
of these packages to unrelated OSGi bundles running inside the
same OSGi framework instance and solves the problem of inconsistent
classpath described in bug 403243 and similar.

Change-Id: I689b89b172d8abe2a733c3e4742e358f27c47443
Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>