Bug 559797 - Delegate to class file editor when no source is found
When no source files are found the source lookup director will not
deletegate the class file editor. This is only happens when the
MavenSourcePathProvider is used. The root cause is the
MavenSourcePathProvider doesn't include the classpath for all the
classes of the dependencies.

This is now fixed by letting the MavenRuntimeClasspathProvider to
provide project classpath entries by removing addProjectEntries from
MavenSourcePathProvider

Change-Id: I39b861c9d329561350b5ce33863bfba9ed85b119
Signed-off-by: gayanper <gayanper@gmail.com>