Create development workspace target platform.
This fixes the build target platform by adding reference to some missing
Maven artifacts.
This also introduce a new dev-workspace target-platform; which is the
same content as build time TP, but augmented with the m2e-maven-runtime
content that usually don't need to be touched and are tricky to deal
with.
The CONTRIBUTING.md got improve to better explain the open and available
target-platform procedures.