[m2e.core] generate modello code into "src-gen" folder
For a plug-in project using a "src-gen" folder is more common. This
change first avoids the creation of a "target"-folder in the project
root by the Maven builder and second avoids that a folder named
"target/generated-sources/modello" is used as source folder (which
is a bit odd because when reading "target" one usually thinks of a
folder that contains build artifacts).

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>