[m2e-core] fix tests failing on Windows
- MavenBugsTests failed on windows because the Inputstream that reads
the content of the modified pom.xml was not closed (it is now replaced
by Files.readString() to avoid stream-handling entirely). As a
consequence an exception was thrown when that pom.xml was deleted at the
end of the test and the test failed.

- WTPEditorTest failed because the cursor was placed hard coded at
location 359 in the pom.xml file's character sequence. Because on
Windows usually two characters are used as line delimiters ("\r\n")
compared to one character on Linux/Mac ("\n") the text at the hard-coded
location is different (it was the middle of the closing '</version>'
tag). The hard code location is replaced by a more robust logic (cursor
is placed after the text of the artifact's version element).