[m2e.core] avoid creation of target-folder by Maven builder
When the m2e.core project is build in a Eclipse-workspace the
maven-builder creates the target-folder containing (nested) only the
messages.properties file.
This is caused by the additional resources-section in the m2e.core/pom.
Since this section is not required with tycho-pomless it is removed in
order to avoid the creation of the target-folder every time te m2e-core
project is build in the IDE.