[Setup] minor improvement of Oomph-setup and adjust CONTRIBUTING.md
The m2e-Oomph-setup is improved in the following way:
- add the variable eclipse.target.platform=None so it is not prompted by
the installer (it is unused).
- Swap order of Launch of m2e-maven-runtime-installion and
refresh/clean/build all. This way the maven-runtime installation is done
before all projects are build, which can avoid a manual refresh/rebuild
or restart.

In the CONTRIBUTING.md some names are corrected, settings of SSH keys
are mentioned and the text is improved in general.

Additionally remove the not intended second 'install' argument of the
m2e-maven-runtime--install launch.

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>