[Setup] perform Maven-runtime installation before TP resolution
Since the Maven-runtime bundles are included via the target-platform, it
is necessary to install them before the target-platform is resolved!
The m2e Maven-runtime bundles are not published to Maven-central and the
intention is to obtain them from the local .m2 repository. But of course
this this only works if they are installed into the local .m2 repository
first.

Additionally the Maven-runtime installation now runs on every start-up
to ensure that developers always have the latest Maven-runtime bundles
installed.

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>