Let IMaven implement Maven
Actually m2e is some kind of "maven" and like the "DefaultMaven" of
maven cli should handle the loading of extensions and alike.

This let IMaven implement Maven and perform some task like the maven CLI
does.
Beside this, a wrapper around a monitor is added that redirects
ExecutionEvents to the monitor.