Fix: Issue 743 - Occasionally, m2e is unable to read wrapped artifacts
When opening e.g. the Target Editor, Eclipse acquires a lock on the
rebundled jar. If m2e then decides that those jars need to be refreshed,
it may corrupt them while updating the Manifest.

Instead, a new, (temporary) file should be created. The name of the new
file is derived from the original jar.

Example:
lucene-analysis-common.jar -> lucene-analysis-common-0.jar

Both the temporary file, as well as the (outdated) original file are
marked for deletion before shutting down the JVM, in order to ensure
that the Eclipse has released the lock again.