Exceptions when failing to wrap Maven artifacts should be logged.
Errors are only shown in the target editor. However, this is a volatile
medium and messages may disappear, when reloading the target platform or
after closing and re-opening the editor.

Writing the error also to the Error Log gives access to the stack trace,
making it easier to find the cause of the problem, in case something
goes wrong.