Add a MavenToolbox
Currently we have some scattered places where we put useful "shorthands"
that perform some maven specific tasks as it is cumbersome to write them
down each time. Often they are not really related to the object itself
but could be written as a function of an execution and/or session.

Because it is often arguable if we should add it to the given interface
to not bloat the API and on the other hand it might became useful on
other places as well, this change introduces a "MavenToolbox" that could
be a place for such methods, easily constructed by a static method but
not intended to be implemented by clients.

This will solve several issues:
- do not require to do the same thing on different places just to not
introduce API
- prevent code from referencing internal API
- can enhance this as we like
- can deprecate methods if we think they are no longer useful without
harming actual m2e implementations