Add xml.resolver to M2E-launch and disable connectors in M2E-setup
Due to an issue with PDE the 'org.apache.xml.resolver' plug-in is not
resolved and therefore not automatically added to the launch. In order
to avoid failures when launching the Eclipse-M2E-IDE temporary add that
plug-in.

And disable M2E connectors in the setup that currently cannot be
installed with the M2E snapshot because they are not yet ready for
m2e-2.0.0