Adjust M2E_VERSION extraction in Jenkinsfile to m2e.sdk/feature.xml
If one modifies the version of the m2e.sdk feature via the PDE-Feature
editor the XML-element 'feature' gets its attributes always written in a
fixed order. But the Snapshot-deployment in the Jenkinsfile relies on
the attribute order to extract the current M2E-Version.
In order to not require modifications in the source tab, this adjusts
the version extraction logic in the Jenkinsfile to the default feature
attribute order.