Merge SLF4J-binding slf4j-simple into o.e.m2e.maven.runtime bundle
The purpose of having the slf4j-simple binding in the separated
o.e.m2e.maven.runtime.slf4j.simple bundle is to make it inaccessible
for the OSGi runtime of the Eclipse-IDE, to not use it when the embedded
Maven-runtime is executed within the IDE. In that case the SLF4J-binding
provided by Eclipse shall be used. Therefore
o.e.m2e.maven.runtime.slf4j.simple simply did not export any packages
making it completely inactive in OSGi.

However when a Maven build is launched from the IDE the embedded
slf4j-simple binding is supposed to be used. This worked because the
Maven-runtime is not a OSGi-runtime and does not care about
Export-Package headers.

The new approach is to embed the jar of the SL4J-binding directly in the
o.e.m2e.maven.runtime bundle and is made invisible for OSGi by simply
not placing it on the Bundle-ClassPath.
The MavenEmbeddedRuntime is adjusted to nevertheless add the
SL4J-binding's jar on the classpath of a launched Maven-build runtime.
To simplify this the 'M2E-SLF4JBinding' header is automatically
generated, whose value is the path to the SLF4J-binding jar within the
o.e.m2e.maven.runtime bundle.