Add full dependencies-closure to classpath of launched embedded Maven
The Eclipse-Platform changed to use Google-Guava jar from
Maven-Central for the 2022-09 release. In contrast to the Guava variant
from Eclipse-Orbit, which was used before, that bundle is not
self-contained and requires classes in other jars like
guava.failureaccess.
Consequently it is not sufficient anymore to only put first-level
requirements on the classpath of the launched embedded Maven runtime. It
has to be the full requirements closure now.