Make archetype-project creation more robust
Use a separate working-directory for archetype-project creation and
ensure that it is used as the maven.multiModuleProjectDirectory.
The parents of the working-directory are searched for a .mvn-directory
and if one is found, it is used as maven.multiModuleProjectDirectory.

Consequently if one creates an archetype-project nested inside an
existing multimodule-project the parent that contains the .mvn-folder is
used as multiModuleProjectDirectory of the archetype-project creation
build. This has the consequences that the maven-extensions defined in
that mmPDir participate in the archetype build.
This is unnecessary and can cause problems if those extensions are
loaded from a repository that is specified in a specified maven-settings
file. This happens for example (and causes build failures) when we test
Tycho-snapshots from the tycho-snapshot repo.

And add printout of archetype-project creation build in case of failure.