Extract referenced slf4j bundle for Maven runtime classpath again
This is necessary because even the slf4j-bundle from Maven-Central is
jar-singed as part of the m2e build when the m2e-repo is assembled.

If the slf4j-jar is not extracted launching the embedded maven-runtime
will fail with errors like:
```
class "org.slf4j.MavenSlf4jFriend"'s signer information does not match
signer information of other classes in the same package
```
The reason for such an error is that the 'slf4j-api' jar and the
'maven-slf4j-provider' jar both provide classes for the package
'org.slf4j'. But all jars that provide classes for the same package must
have the same jar-signer. And while the 'slf4j-api' jar is referenced as
bundle and therefore signed in the m2e build (although it is originated
from Maven-Central), the 'maven-slf4j-provider' jar embedded into the
m2e.maven.runtime is not jar signed (because nested jars are not
signed).
Extracting the jar and providing them from a directory
This check is not performed if a jar is extracted and added to the
classpath as directory. Therefore only the slf4j jar is extracted again.
Effectively this reverts commit 143c182b788e0f5e37dc36c63b24a13d2279ccc0
for slf4j.

Additionally this ensures that paths to jars are canonicalized.