Deprecate (and remove usages) of IMaven.readMavenProject
IMaven.readMavenProject(File, ProjectBuildingRequest) is dangerous to
use in global context (aka IMaven) as it do not take basedir /
extensions into account. Actually reading the raw maven model should not
be exposed to consumers at all as it is very easy to get wrong and do
not support caching.

This deprecates the method and converts usages to internal toolbox
instead that are used in the correct context.