Adjust MavenExecutionContext to be more well-defined
Currently we do some computations regarding the basedir in
MavenExecutionContext that later influences the session created, but
actually this should be the concern of the creator to decide if and what
basedir has to be used, e.g. even if there is currently no project a
basedir might still be known.

This changes the MavenExecutionContext to accept a basedir on
construction time and changes the ImavenFacade to a more generic
supplier.