Clear reference to the ProjectBuildingRequest after loading the project
Currently a MavenProject holds a reference to the ProjectBuildingRequest
(which is big) and thus retains the object in memory as long as the
MavenProject is in memory.

The method itself is deprecated in maven and explained as

"This is an utility method that is meant to assist integrators of Maven,
it must not be used by Maven plugins."

As m2e do not used this method it seems valid to clear the reference
always to conserve space.