Optimize Caching by distributing already loaded maven projects
Currently when a maven project is loaded all its parent projects are
loaded as well. For deeply nested multi-module setups this can easily
lead to filling the cache with hundred of duplicate projects. This
reads maven projects multiple times and even worse reduces an effective
caching as we need to limit the cache to a smaller size with worse
performance even though the data is still retained in memory.

This changes the follwoing things:

1) do cache on the ArtifactKey and path of a pom instead of the facade
object itself to always ever keep one instance in the cache per maven
facade identity
2) distribute loaded parent projects to the cache so these not need to
be loaded again if requested afterwards
3) raises the default caching size from 20 to 50 to account for more
efficient caching