Add a Quick-Fix to add a dependency on missing types
Currently if a type is missing a user needs to search the workspace,
then probably find the project and then go to the pom and add the
desired coordinates.

This now adds a Quick-Fix to automate this and offers the user to add a
dependency to a workspace project that contains the missing type.