Add new IProjectConfiguration as replacement for ResolverConfiguration
Currently ResolverConfiguration is used for many purpose especially for
grouping projects in resolve operations. The problem is that this class
itself is modifiable and thus can change without any notice and as it is
a public API class we can't easily change this or have alternative
implementations. Beside that the javadoc suggest that the naming is not
well chosen.

This adds a new interface IProjectConfiguration that also takes the new
concept of MultiModuleProjectDirectory into account