Execute Maven plug-ins to generate sources required for M2E compilation
The connectors for the 'maven-dependency-plugin' and the
'modello-maven-plugin' are currently broken because they are not yet
adapted to M2E 2.0. Threrefore command m2e explicitly to execute those
goals.

In order to make this work in the IDE the o.e.m2e.core/pom.xml had to be
re-created.