[Build] Perform p2-baseline-replace for maven-runtime too
and use baselineReplace=all instead of 'common'. Common was only used
when the source bundles for the maven-runtime were newline introduced to
not have them removed. But since that is established now 'all' can be
used again.

Moving the tycho-source-plugin to the parent also allows to execute the
eclipse-jarsigner-plugin:sign goal in its default (package) phase. This
is caused by the way Maven merges executions from parent, childs and
profiles.