[TP] update to Eclipse 4.26 release and latest versions of other deps
Additionally update M2E's launch configs to new dependencies and to use
new attributes.

The qualifier of the 'org.eclipse.m2e.model.edit' Plugin has to be
increased because somehow the resulting byte-code of the classes in the
org.eclipse.m2e.model.edit.pom.translators has changed with this update.