[Build] Temporarily ignore baseline discrepancy for maven-runtime
... and bump qualifiers for changed projects.

Due to changes in Tycho the metadata of the maven-runtime component's
sources are now different when compared on textual level, but for
consumers that change is not relevant and therefore the difference can
be ignored.

Also enforce qualifier updates of some Plugins the compiled byte-code
has changed with this update. Probably due to newer JDK minor version
used in the CI.