Do not wrap source bundles
Currently m2e wraps source into new artifacts having performance and
space implications and can be confusing because users expect the file to
be on a different location.

Due to the recent changes in PDE this is no longer required an we can
remove this workaround.

See https://github.com/eclipse-pde/eclipse.pde/pull/425