[Build] Move args to maven.cfg and use p2.baselineMode=failCommon in CI
Move common Maven args that should be used in the CI and for local
builds to the maven.config.
Use the tycho.p2.baselineMode=failCommon in Jenkins and in GitHub
workflows, but in general to be more gentle in local builds. Move its
definition out of the 'eclipse-sign' profile.

And prevent Maven build launches of AbstractMavenProjectTestCase from
using the .mvn folder (and its content) from this git-repo.