Don't open the dialog if we can't propose anything to the user
Currently if not a single connector can be proposed the dialog is still
opened and disturbs the user even though there is nothing much one can
do (and most probably want to do).

This first checks if there is at least one proposal before open the
dialog when importing a project.

This relates to https://github.com/eclipse-m2e/m2e-core/issues/1012