Execute unknown mojos by default
Currently unknown mojos are marked as an error and are ignored, users
get a warning and no incremental builds are performed.

This is now changed to always execute mojos in incremental builds and
adding an option for the user to choose the default action to take.

Relates to #1013
Apply suggestions from code review

Co-authored-by: Hannes Wellmann <wellmann.hannes1@gmx.net>