Initialize the dependency table viewer with SWT.FULL_SELECTION.
This is required for tables with more than one column, where each column
should be editable. If you want to edit a cell, you currently have to
first select the row you want to edit, by clicking on the item in the
first column. In order to then edit a different cell, you then have to
first select the new column, before the cell becomes editable.

This issue only seems to exist on Windows machines.