Don't show the DiscoveryWizard if nothing new was discovered
Currently we show a DiscoveryWizard after project import, even if no new
items are discovered. This is not really helpful for the user, as the
only thing one can do is to completely ignore an execution (what is
always possible from the pom afterwards) or cancel the dialog.

This changes the project import to not show the dialog if nothing was
found, but shows the full dialog if the user manually triggers the
discovery.