Do not "warn" on incomplete mappings
M2E has had a dialog that warns the user if there are "incomplete"
mappings, but actually the user can't do anything useful. Also m2e for a
while handles missing mappings with reduced severity (e.g. by just add a
warning marker) and now even executes them by default.

Because of this, the preference an dialog has even lesser value and only
disturbs the user and is therefore obsolete.