Improve wording and icons in the DiscoveryDialog
Currently the DiscoveryDialog shows missing messages with a big red x
and a text that says "resolve later", this gives an expression something
is generally broken / error. Actually m2e now executes all mojos by
default, and even before it shows non executed mojos as warnings.

This replaces the error icon with a light-bulb/info icon so the user
knows there is something one can take an action on but it is probably
not that important to be concerned about (right now), and adjust the
text to reflect that if nothing is choose it will just do whatever is
choose in the workspace settings.