Only trigger (incremental) builds if there are relevant changes
Currently any file change in a project can trigger a rebuild of that
project, even if it is from the build-output or from child modules. This
can produces "endless" loops of the build and produces a lot of build
rush if automatic refresh is activated for the workspace or resources
are refreshed on access. On the other hand, without autorefresh some
changes are never detected.

This changes the following:
- check if a delta only contains irrelevant deltas and then do not
trigger an incremental maven build
- refresh the whole project first to discover new files generated
- if auto refresh is enabled, do not manually refresh but use the native
refresh events

This can considerably reduce the number of rebuilds happening.

Fix https://github.com/eclipse-m2e/m2e-core/issues/1275