[Maven-Runtime] Support slf4j.api version 2 in m2e.maven.runtime
As part of the new SLF4J-2 major version the mechanism to connect the
slf4j.api with a logging back-end has been changed from the
static-binding mechanism to Java's ServiceLoader mechanism [1].

In order to support slf4j.api-2 a provider for the maven-slf4j-provider
(the logging backend provided/used by Maven based on slf4j-simple) is
needed. Because maven-slf4j-provider does not provide it itself we have
to supply it for m2e.maven.runtime.

[1] - https://www.slf4j.org/faq.html#changesInVersion200