Compute the full dependency chain for wrapped artifacts
Currently when wrapping a jar we only add this one jar to the analyzer,
this has the problem that the Analyzer has no clue about the used
dependencies resulting in poor metadata, e.g. missing package-import
version ranges.

This introduces a new MavenBundleWrapper that computes the full
dependency chain for a jar to be wrapped (possibly wrapping them as
well) and maintains a persistent id of the instructions so that wrapped
items can be reused across invocations and tools.