Do not embed guice in the runtime bundle
google guice is already a bundle, it should therefore not be required to
embed it into the m2e.runtime, especially as we then export a package
again.

This will reduce the size of m2e runtime, allows reuse of exiting guice
bundles and prevents us from the need to bump the bundle less often in
case of updates.