Make the target test plugable so they can be reused (e.g. by Tycho)
One of the important parts is that Tycho+m2e behave the same in regard
to the m2e target support so that if something works in the IDE it also
works in the build.

This is an attempt to generlaize the tests in a way so they can be
reused at Tycho by providing a small loader that transform the Tycho
content into something understood by the test, so ideally we can run the
same tests in Tycho as in m2e.

Signed-off-by: Christoph Läubrich <laeubi@laeubi-soft.de>