Add a preference to configure the relevance and search for workspace QuickFix
This adds a preference for the relevance of a quickfix, due to how
quickfixes work there is no real way to determine how important one is,
it is only possible to give an absolute number. To allow user to
higher/lower the quickfix of m2e it is now possible to configure the
value.

Beside that it is possible to configure if the quickfix for searching
workspace projects for missing types is enabled/disabled.