Do not prevent finish for "incomplete" mappings
Currently in #finish of the wizard it can happen that the wizard do not
finish if m2e thinks the mapping is incomplete, but this is a very bad
user experience as there is not indication of what has to be done.

Instead a page should set its completion status (in this case the button
is disabled) and instruct the user what to do instead.

Fix https://github.com/eclipse-m2e/m2e-core/issues/1371