Consume dependencies of maven archetype from the target platform
Currently we use a special packed "uber-bundle" that contains everything
as library entries. This makes the installation and building quite
complicated as we need to take special care regarding making the source
available. Also upgrading individual items is not possible without
version bumps, on the other hand the m2e-target support already offers
all we need here and can even generate a feature out of the selected
items.

This changes the structure in the following way:
- mark the m2e archetype bundle with extensible-api header
- add all dependencies of maven archetype as fragment to the target
- remove the embedded jars