Add support for multiple embedded maven runtimes
Currently m2e only offers one embedded runtime, and for all other cases
the user has to setup a local install, but actually the embedded runtime
bundle is independently installable so m2eclispe can offer more than one
maven runtime to be used for launches.

This makes the required changes to support multiple runtime bundles to
be selected by the user from the UI and use in launches.