Retain Maven 3.8 compatibility
m2e can perfectly build with maven 3.8 runtime so we should retain this
as long as possible to have a way for people to switch to older runtimes
but using newer m2eclipse.

The only issue is that we need to use a deprecated class here but unless
it is really removed from maven we can simply use that without an issue.