Abstract the delta handling from the resource API
Currently m2e is directly wired to the resource delta API, while this
seems good in the context of eclipse for m2e this imposes to restrictive
API as it is all noimplement/extend interfaces and offer much more than
m2e actually using.

This aims to implement a thin wrapper for further improvement of delta
detection especially when it comes to resource see for example
https://github.com/eclipse-m2e/m2e-core/issues/1302

This will not fix anything but is an important first step to ensure
things are still working as before.