Record the build timestamp and assume every file changed that is newer
Currently we only analyze the current delta but this can have several
subtile problems:

- when a file is modified outside of eclipse it might not be recognized
by eclipse immediately
- when files are modified outside the "interesting folders" the delta
maybe not includes this
- if a build fails the next delta will not necessarily include all files
of the previous delta because the error might be in a java source file
and a resource file was changed

This adds a project scoped timestamp to record when the last successful
build happens and assumes every file changed that has a timestamp larger
than the last build, this still does not handle deletes properly but
improves the situation already for existing files. It also leads to
probably more files queried/assumed changed in case of failing builds or
when the build timestamp is unkown but ensures that we only ever process
more files than required but never less ones.