Change Micro-Version schema in wrapped archetype artifacts to 100-steps
Before the Maven-archetype artifacts where obtained and wrapped in the
TP, when they were embedded in a single 'o.e.m2e.archetype.common'
bundle, said bundle had the version 3.2.104 and used 100er step as micro
version where the hundreds matched the version of the original maven
artifact's version.
This schema should be preserved for now to not have back-steps in
versions until the next minor of Maven archetype is available.