MavenTargetLocation: use serialize for equals and hashCode
In the TargetPlatformPreferencePage, upon the Apply button, some target
definition containing some MavenTargetLocation always was rewritten to
it's .target file. This was because MavenTargetLocation.equals never
returned true for multiple reasons, starting with MavenTargetDependency
lacking an impl of equals().
Using serialize(), we can be sure to perform the equals check on any
relevant attribute, including deeply nested ones.

Beyond that, also use serialize() to implement hashCode().