Support project local maven settings
Currently m2e has one global way to specify the global/user settings
file, but one can also configure this through .mvn/maven.config files
and while m2e already supports specification of properties and profiles
there this does not work for (global)settings. This leads to the
situation that user probably getting different results when running a
mojo as part of eclipse or on the commandline.

This now adds the infrastructure and support for projects having a local
configuration as well by augmenting the maven.config if available.