Add testcase for classifier with transitive dependencies
Currently if a dependency is specified  the transitive items seem to be
not correctly collected, at laest there is one artifact (ehcache) that
has a transitive dependency that is invalid (javax.xml) and one can
select a different classifier (jakarta) to use the jakarta ones but this
seem not to happen at the moment.

This adds a testcase to illustrate the problem.