Trace test execution directories in the remote maven execution
Currently running tests in maven via m2e is possible but not very
convenient as one still needs to navigate to the results then open the
correct file. Another pitfall is that even if one has opened the file
once, the "classic" JUnit view not update the view even when the file
changes afterwards.

This now adds a new process tracking of test executions directories that
then can be watched on the m2e side and display the new advanced JUnit
view when the run has finished.