Make the remove maven listener a process in the launch
Currently m2e uses an additional socket connection to receive project
events from the remote maven project. This is not visible to the user
and requires some internal maps to maintain state.

This refactors this into a more generic event facility and make it
visible as a Process in the debug view, this also has the advantage that
the process itself can be used to maintain the state.