Revert "Update plexus-build-api to codehaus one"
The plexus-build-api [1] can be implemented by Maven Mojos in order to
support IDE's (like Eclipse+M2E) to improve their integration with the
implementing Mojo.

But just embedding the new API jar without adding support for it in M2E
does not have any effect and just increases the size of the runtime.
Therefore this should be reverted and re-added together with the
necessary changes in M2E's code to support it.
That being said it would not be harmful to just update the dependency
because org.codehaus.plexus:plexus-build-api:1.2.0 has a dependency to
org.sonatype.plexus:plexus-build-api:0.0.7 and therefore also pulls it
into the m2e.maven.runtime.

This reverts commit 847581dbb7879e65163c304bac03ba0b004a278f.

[1] - https://github.com/codehaus-plexus/plexus-build-api