[Maven Runtime] Reduce number of manually maintained version
1. Remove manually maintained version of exported packages
The bnd-maven-plugin can derive the version of the org.eclipse.aether*
and org.eclipse.sisu* packages perfectly fine based on the OSGi metadata
that are embedded in these jars (since some time).

Furthermore the version of the exported org.eclipse.sisu* packages has
been false before (due to the false version being used manually).

2. Exclude the dependency to 'commons-codec' explicitly instead of
mentioning it as provided to exclude it from the runtime bundle.
Excluding it does not require a version to be defined.