Perform update of Maven artifacts in non-UI thread
Updating the selected artifacts in the target dependency dialog might a
long-running operation, as a request needs to be sent to Maven Central,
and should therefore run without blocking the application.

Instead of being executed in the UI thread, this is now done in the
runnable context of the containing wizard.