Fix transitive artifact collection in m2e targets and make sharable
Currently m2e uses the usual maven resolver to find transitive
dependencies but this works differently than what we want here. Instead
we want all dependencies and their dependencies and then filter for the
given scopes provided if any match.As this is a non trivial operation,
it seems useful to make the code also sharable with Tycho so we ensure
there is always the same result in Tycho as in m2e.

This now creates a new MavenDependencyCollector that can be shared with
Tycho that traverses the dependency tree in a way we need here and
collects everything along the way. This also has the advantage, that if
we choose to only collect direct dependencies, not the whole tree has to
be traversed as before.