Use Maven 4 as the runtime for m2eclipse
Maven 4 will soon be released and should be able to execute Maven 3
projects as well. We should therefore upgrade m2e to use Maven 4 from
now on.

For users that require Maven 3.9 line we can offer the latest 3.9.11
maven runtime as an additional install option like we already did for
Maven 3.9.6